% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sp1}
\alias{sp1}
\title{Example Data: Conium maculatum's native distribution}
\format{An data frame with 4975 rows and 3 variables:
\describe{
  \item{sp}{species name}
  \item{x}{longitude}
  \item{y}{latittude}
}}
\usage{
data(sp1)
}
\description{
Example Data: Conium maculatum's native distribution
}
\examples{

library(humboldt)

##load environmental variables for all sites of the study area 1 (env1). Column names should be x,y,X1,X2,...,Xn)
data(env1)
## load environmental variables for all sites of the study area 2 (env2). Column names should be x,y,X1,X2,...,Xn)
data(env2)

## remove NAs and make sure all variables are imported as numbers
env1<-humboldt.scrub.env(env1)
env2<-humboldt.scrub.env(env2)

##load occurrence sites for the species at study area 1 (env1). Column names should be sp,x,y
occ.sp1<-data(sp1)

##load occurrence sites for the species at study area 2 (env2). Column names should be sp,x,y
occ.sp2<-data(sp2)
}
\keyword{datasets}
