% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humboldt79.R
\name{background.iter.r}
\alias{background.iter.r}
\title{internal fuction}
\usage{
background.iter.r(
  repi,
  z1,
  scores.env12a,
  scores.env2a,
  scores.sp2a,
  scores.env2c,
  Rana.x,
  Rana.y,
  env.reso.sim,
  run.silent.bak,
  inE,
  Rin,
  ncores,
  userOS,
  threshinZ,
  kern.smoothinZ,
  pb,
  repT
)
}
\description{
internal fuction
}
\keyword{internal}
