% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humboldt85.R
\name{equivalency.iter}
\alias{equivalency.iter}
\title{internal fuction}
\usage{
equivalency.iter(
  repi,
  z1_sp,
  z1_glob,
  z2_sp,
  z2_glob,
  nacinZ,
  threshinZ,
  kern.smoothinZ,
  Rin,
  pb,
  repT,
  ncores,
  run.silent.equ,
  cor.envn,
  userOS
)
}
\description{
internal fuction
}
\keyword{internal}
