% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humboldt60.R
\name{equivalency.iter}
\alias{equivalency.iter}
\title{internal fuction}
\usage{
equivalency.iter(repi, z1, z2, nacinZ, threshinZ, kern.smoothinZ, Rin, pb,
  repT, ncores, run.silent.equ, cor.env.n, userOS)
}
\description{
internal fuction
}
\keyword{internal}
