% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humboldt49.R
\name{humboldt.doitall}
\alias{humboldt.doitall}
\title{'Do it all'. Performs most analyses in Humboldt and outputs the results as Jason likes!}
\usage{
humboldt.doitall(inname = "DoItAll", env1, env2, sp1, sp2, rarefy.dist = 0,
  rarefy.units = "dd", env.reso, reduce.env = 2, reductype = "PCA",
  non.analogous.environments = "NO", nae.window = 5, env.trim.mcp = T,
  mcp.buffer.sp1 = 200, mcp.buffer.sp2 = 200, correct.env = T,
  col.env = e.var, e.var, R = 100, kern.smooth = 1, PROJ = F,
  e.reps = 100, e.sample.method = "points", b.reps = 100,
  b.force.equal.sample = F, env.pca = T, nae = "YES",
  thresh.espace.z = 0.001, p.overlap = T, p.boxplot = F, p.scatter = F,
  run.silent = F)
}
\arguments{
\item{inname}{string for labelling results output from function}

\item{env1}{environmental variables for all sites of the study area 1 (env1). Column names should be x,y,X1,X2,...,Xn; with X1-Xn being any string label. If env1=env2, input the same file twice.}

\item{env2}{environmental variables for all sites of the study area 2 (env2). Column names should be x,y,X1,X2,...,Xn; with X1-Xn being any string label. If env1=env2, input the same file twice.}

\item{sp1}{occurrence sites for the species/population 1 at study area 1 (env1). Column names should be 'sp', 'x','y'}

\item{sp2}{occurrence sites for the species/population 2 at study area 2 (env2). Column names should be 'sp', 'x', 'y'}

\item{rarefy.dist}{remove occurrences closer than a minimum distance to each other (this function uses the humboldt.occ.rarefy function). Values need to be in km[recommended] or decimal degrees. See associated parameter rarefy.units. Note: rarefy.dist=0 will remove no occurrences}

\item{rarefy.units}{the units of rarefy.dist parameter, either "km" for kilometers or "dd" for decimal degrees}

\item{env.reso}{the resolution of the input environmental data grid in decimal degrees}

\item{reduce.env}{the format to trim environmental space so that it is shared. If reduce.env=1, the second input environment (env2) will be trim the match the first input (env1). If reduce.env=2, both input environments trimmed so that extents of both are identical (the lower maximum value observed in env1 and env2 and the higher minimum value observed in env1 and env2 will be used to trim environmental space for each PC/environmental variable) If reduce.env=0, you will skip trimming environmental space}

\item{reductype}{if reduce.env= 1 or 2, the 'reducetype' parameter specifies the format for how to reduce environmental space ("PCA" or "STANDARD"). If reductype="PCA", the environmental space will be trimmed based on two principal components. If reductype="STANDARD", the environmental space will be trimmed by each included variable specified in col.env. If reduce.env=0, do not include this parameter}

\item{non.analogous.environments}{allow non-analogous environments in environmental space? If non.analogous.environments="YES", non-analogous environments between env1 and env2 will retained. If non.analogous.environments="NO", non-analogous environments between env1 and env2 will be removed. This parameter is only usable under the combinations of reductype="PCA" & reduce.env=1 or reductype="PCA" & reduce.env=2. If something else, this parameter will not be used.}

\item{nae.window}{the spatial window from which non-analogous environments will be quantified.  The non-analogous environments are characterized by gridding the espace of env1 and env2 into a R x R grind (e.g. 100 x 100).  If nae.window=0, values absent from a cell in one environment will be removed from the other.   If nae.window>0, values absent from a window (or neighborhood of cells) in one environment will be trimmed from the other.  The nae.window value characterizes the number of cells to search from the focal cell of environmental space values in the other environment.  The larger the nae.window value, the fewer non-analogous environments removed. This parameter allows imperfect overlap of environments.   If areas of environmental space are a little patchy between environments—but generally present-- a larger nae.window value will retain more of the patch environments. The default value is a nae.window=5}

\item{env.trim.mcp}{trim extent of environmental data by buffered MCP in geographic space. Necessary for comparing different species or different extents and for removing non-accesible environments}

\item{mcp.buffer.sp1}{buffer distance (in km) for trimming available environmental space for sp1}

\item{mcp.buffer.sp2}{buffer distance (in km) for trimming available environmental space for sp2}

\item{correct.env}{if correct.env=T, the analysis corrects occurrence densities of each species by the prevalence of the environments in their range. If correct.env=F, the overlap measure does not correct occurrence densities of each species by the prevalence of the environments in their range.Default=T}

\item{col.env}{if reductype="STANDARD", then parameter specifies the number of columns to trim environmental space on. This can be any number of columns}

\item{e.var}{selection of variables to include in all of the analyses- this is a separate parameter than col.env but must contain all variables included in col.env}

\item{R}{resolution of grid in environmental space (RxR)}

\item{kern.smooth}{scale at which kernel smoothing occurs on environmental data, larger values (i.e. 2) increase scale (making espace transitions smoother and typically larger) and smaller values (i.e. 0.5) decrease scale (making occupied espace clusters more dense and irregular). Default value is 1.  You can also input: "auto", which estimates the kernel parameter by calculating the standard deviation of rescaled PC1 and PC2 coordinates divided by the sixth root of the number of locations. This method can be unreliable when used on multimodal espace distributions as it results in over-smoothing of home ranges.  Multimodal espace occupancy can be somewhat common when a species occupies an extreme aspect of habitat or when espace is not broadly accessible in both dimensions of espace (PCs 1 & 2)}

\item{PROJ}{If PROJ =F, the models are calibrated on both ranges. If PROJ =T, the models are calibrated on species 1 range only and projected to range 2}

\item{e.reps}{the number of iterations for the equivalency test (humboldt.equivalency.test). Values higher than 200 are recommend for final analysis}

\item{e.sample.method}{this parameter controls how espace is sampled.  There are two options: "points" or "density".  If e.sample.method="points", only input localities for both species are used for randomizations. If e.sample.method="density", the gridded density of species' espace is randomly resampled with higher sampling probabilities applied to areas of high density espace.  In both cases the number of samples selected for each species is equal to the observed datasets.  Again 'points' focuses only on input points, whereas 'density' samples the entire gridded density of each species' espace. Generally I prefer using 'points' for most datasets, because it maintains most of the nuances associated the original datasets (i.e. spatial autocorrelation).  Though sometimes I do use 'density' for datasets with limited localities (<20 points)}

\item{b.reps}{the number of iterations for the Background tests (humboldt.background.test). Values higher than 200 are recommend for final analyses}

\item{env.pca}{calibrate PCA on only occurrence data (env.pca=F) or environment data (env.pca=T), recommended is env.pca=T}

\item{nae}{do you include non-analogous environments in the niche overlap measurement? If nae="NO" (use captial letters), then non-analogous environments will be removed from both input environments during overlap measurement and only environments present in both datasets will be used to measure overlap. If nae="YES", then no change will be made to input z1 and z2. Note: this is separate from trimming non-analogous environments from your input dataset (as done by humbodlt.g2 specified by parameter non.analogous.environments). This parameter physically removes non-analogous environments from datasets ONLY before the niche overlap measurement. Technically the removal of non-analogous environments via either way should result in similar overlap measurements (though they may not be identical). This because removing NAE from the dataset prior to gridding environments will resulting only non-analogous environments to be gridded (and typically finer grain applied to each grid cell). Whereas removing them only via this parameter (nae), which only removes non-analogous in the gridded environmental space for use in overlap measurements--- all the input environmental space is gridded (likely increasing the environmental space per gridded cell). A second cause of differences in values can result from rescaling of espace values during niche-overlap measurements so that the sum of the landscape equals one. If occupied non-analogous environmental are numerous in one of the datasets, this can theoretically cause overlap values to decrease in analogous environments (vs. nae) because differences in core niches are rescaled to 1 in both scenarios. The rescaling among fewer cells increases the values applied to highly suitable areas and, if not equivalently scaled in both datasets, differences among niches could increase, resulting a smaller overlap in non-analogous environments (again values should be similar). If you remove non-analogous environments in humboldt.g2e, I also suggest that you use this function (as it can remove any slight anomalies caused by gridding environments in humbodlt.grid.clim due to the binning of values in the RxR grid).}

\item{thresh.espace.z}{this parameter is an experimental parameter and controls the level at which values below the kernel density z values are removed for creating areas of analogous environmental space. Higher values will increase value from which the low-density areas are removed from the environmental space of z1 and z2.  Basical values above this are retained and values below are removed. Default=0.001}

\item{p.overlap}{if p.overlap=T, the niche overlap plot will be output (humboldt.plot.overlap). Turn this to 'F' if you want to speed of analyses. This plot takes a lot of CPU time because it calculates kernel densities for all environment and species datasets input}

\item{p.boxplot}{if p.boxplot=T, a boxplot niche overlap plot will be created. The whisker of boxplot depict the environmental space of environment. Whereas the bars depict environmental space of each species within that environment. Dots on bars depict density of species localities in environmental space. This is a quick plot and doesn't requires lots of CPU time}

\item{run.silent}{if run.silent=T, texts boxes displaying 'sampling', 'rarefying', 'equivalency test', 'Background test' progress will not be displayed}

\item{p.overlap}{if p.scatter=T, this will plot scatter plots with histrograms of your environment and species datasets using the humboldt.plot.scatter function. This is a quick plot and doesn't requires lots of CPU time}
}
\value{
Performs almost all the analyses in Humboldt and outputs the results as Jason likes! See my example for my recommended running. For each comparison I recommend running the 'doitall' function twice. In the first run, I highly recommend running the analysis on the full environment input for both background and equivalency tests. In this case you are testing the total equivalence between species (or divergence) in their current realized distributions. It asks the question how equivalent are the two species realized niches?  This test is also the best gauge for background test and the power to discriminate differences.  I only use background values from this.  However, I do re-run the analyses with trimmed, shared espace. This is extremely important if you are interested in quantifying niche evolution/divergence among two populations/species . This test how two species diverge in shared analogous e-space and is the only actual test of niche divergence/evolution. Note that non-significant background test are not uncommon in shared-analogous espace and should not be given priority overlap total espace measurements from run 1. Under the situation where there is very little or no analogous e-space ---you may not be able conclude much from this situation regarding niche divergence. In this situation, if the species' 'niches' are not truncated by espace boundaries, then this suggests that the g2e conversion is a decent approximation of the species' fundamental niches. And because of this, the results should be accurate and species divergence can also be concluded. However if the core densities on one or both of species are on margins or near margins of espace, then this suggests that the fundamental niche may be much larger than available climate space provides. In this latter case, you cannot conclude that the species have diverged at all. In the former case, where there is no truncation of espace of each species in their habitats, but little overlap between, this is evidence the species have diverged.  In this case, simply report the niche overlap values in the total environmental space and the amount analogous espace.
}
\description{
'Do it all'. Performs most analyses in Humboldt and outputs the results as Jason likes!
}
\examples{
#######################################################################################
###################################    EXAMPLE 1    ###################################
#######################################################################################
library(humboldt)
##load environmental variables for all sites of the study area 1 (env1). Column names should be x,y,X1,X2,...,Xn)
env1<-read.delim("env1.txt",h=T,sep="\\t")

## load environmental variables for all sites of the study area 2 (env2). Column names should be x,y,X1,X2,...,Xn)
env2<-read.delim("env2.txt",h=T,sep="\\t") 

## remove NAs and make sure all variables are imported as numbers
env1<-humboldt.scrub.env(env1)
env2<-humboldt.scrub.env(env2)

##load occurrence sites for the species at study area 1 (env1). Column names should be sp,x,y
occ.sp1<-na.exclude(read.delim("sp1.txt",h=T,sep="\\t"))

##load occurrence sites for the species at study area 2 (env2). Column names should be sp,x,y 
occ.sp2<-na.exclude(read.delim("sp2.txt",h=T,sep="\\t"))

##its highly recommened that you using the function "humboldt.top.env" to select only the important enviromnetal variables in humbodlt.doitall. This step can be skipped. If you downloaded tons of environmental data, you should use this step.  If you skip this step, input env1/env2 inplace of reduc.vars$env1/reduc.vars$env2 
reduc.vars<- humboldt.top.env(env1=env1,env2=env2,sp1=occ.sp1,sp2=occ.sp2,rarefy.dist=40, rarefy.units="km", env.reso=0.416669,learning.rt1=0.01,learning.rt2=0.01,e.var=(3:21),pa.ratio=4,steps1=50,steps2=50,method="contrib",contrib.greater=5)

##Adjust the number of variables input for e.vars after reduction to only important variables
num.var.e<-ncol(reduc.vars$env1)
##run it first with full environmental for backgroud tests and equivalency test (total equivalence or divergence in current distributions)
full<-humboldt.doitall(inname="full_extent", env1=reduc.vars$env1, env2=reduc.vars$env2, sp1=occ.sp1, sp2=occ.sp2, rarefy.dist=50, rarefy.units="km", env.reso=0.416669, reduce.env=0, reductype="PCA", non.analogous.environments="YES", correct.env=T, env.trim.mcp=F, mcp.buffer.sp1=200, mcp.buffer.sp2=200, col.env=e.var, e.var=c(3:num.var.e), R=100, kern.smooth=1, PROJ=F, e.reps=100, e.sample.method="points", b.reps=100, env.pca=T, nae="YES",thresh.espace.z=0.001, p.overlap=T, p.boxplot=F, p.scatter=F, run.silent=F)

##run it a second time with a trimmed, shared-espace. Here the equivalency test tests for niche evolution or niche divergence. For comparing results, change only the following model parameters: reduce.env, non.analogous.environmental,env.trim.mcp
shared_ae<-humboldt.doitall(inname="shared_espace_ae", env1=reduc.vars$env1, env2=reduc.vars$env2, sp1=occ.sp1, sp2=occ.sp2, rarefy.dist=50, rarefy.units="km", env.reso=0.416669, reduce.env=2, reductype="PCA", non.analogous.environments="NO", correct.env=T, env.trim.mcp=T, mcp.buffer.sp1=200, mcp.buffer.sp2=200, col.env=e.var, e.var=c(3:num.var.e), R=100, kern.smooth=1, PROJ=F, e.reps=100, e.sample.method="points", b.reps=100, env.pca=T, nae="YES",thresh.espace.z=0.001, p.overlap=T, p.boxplot=F, p.scatter=T,run.silent=F)
#######################################################################################
###################################    EXAMPLE 2    ###################################
#######################################################################################
############################  Using Provided Example Data   ###########################
#######################################################################################
library(humboldt)
##load environmental variables for all sites of the study area 1 (env1). Column names should be x,y,X1,X2,...,Xn)
data(env1)
## load environmental variables for all sites of the study area 2 (env2). Column names should be x,y,X1,X2,...,Xn)
data(env2)

## remove NAs and make sure all variables are imported as numbers
env1<-humboldt.scrub.env(env1)
env2<-humboldt.scrub.env(env2)

##load occurrence sites for the species at study area 1 (env1). Column names should be sp,x,y
data(sp1)

##load occurrence sites for the species at study area 2 (env2). Column names should be sp,x,y
data(sp2)

##its highly recommened that you using the function "humboldt.top.env" to select only the important enviromnetal variables in humbodlt.doitall. This step can be skipped. If you downloaded tons of environmental data, you should use this step.  If you skip this step, input env1/env2 inplace of reduc.vars$env1/reduc.vars$env2 
reduc.vars<- humboldt.top.env(env1=env1,env2=env2,sp1=sp1,sp2=sp2,rarefy.dist=50, rarefy.units="km", env.reso=0.416669,learning.rt1=0.01,learning.rt2=0.01,e.var=(3:21),pa.ratio=4,steps1=50,steps2=50,method="contrib",contrib.greater=5)

##Adjust the number of variables input for e.vars after reduction to only important variables
num.var.e<-ncol(reduc.vars$env1)
##run it first with full environmental for backgroud tests and equivalency test (total equivalence or divergence in current distributions)
full<-humboldt.doitall(inname="full_extent", env1=reduc.vars$env1, env2=reduc.vars$env2, sp1=sp1, sp2=sp2, rarefy.dist=50, rarefy.units="km", env.reso=0.416669, reduce.env=0, reductype="PCA", non.analogous.environments="YES", correct.env=T, env.trim.mcp=F, mcp.buffer.sp1=200, mcp.buffer.sp2=200, col.env=e.var, e.var=c(3:num.var.e), R=100, kern.smooth=1, PROJ=F, e.reps=100, e.sample.method="points", b.reps=100, env.pca=T, nae="YES",thresh.espace.z=0.001, p.overlap=T, p.boxplot=F, p.scatter=F, run.silent=F)

##run it a second time with a trimmed, shared-espace. Here the equivalency test tests for niche evolution or niche divergence. For comparing results, change only the following model parameters: reduce.env, non.analogous.environmental,env.trim.mcp
shared_ae<-humboldt.doitall(inname="shared_espace_ae", env1=reduc.vars$env1, env2=reduc.vars$env2, sp1=sp1, sp2=sp2, rarefy.dist=50, rarefy.units="km", env.reso=0.416669, reduce.env=2, reductype="PCA", non.analogous.environments="NO", correct.env=T, env.trim.mcp=T, mcp.buffer.sp1=200, mcp.buffer.sp2=200, col.env=e.var, e.var=c(3:num.var.e), R=100, kern.smooth=1, PROJ=F, e.reps=100, e.sample.method="points", b.reps=100, env.pca=T, nae="YES",thresh.espace.z=0.001, p.overlap=T, p.boxplot=F, p.scatter=T,run.silent=F)
}
\seealso{
\code{humboldt.g2e, humboldt.equivalency.test, humboldt.background.test, humboldt.niche.overlap, humboldt.plot.niche, humboldt.doitall, humboldt.top.env} which use or depend on outputs of this function
}
