% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humboldt47.R
\name{humboldt.plot.contrib}
\alias{humboldt.plot.contrib}
\title{Plot PCA contribution to environmental space}
\usage{
humboldt.plot.contrib(contrib = pca.cal$co, eigen = pca.cal$eig)
}
\arguments{
\item{contrib}{The contribution of the environmental variables included in PCA}

\item{eigen}{Eigen values output from ordinations}
}
\value{
Plots the contribution of the environmental variables to the analysis. Typically these are eigen vectors and eigen values in ordinations.
}
\description{
Plot PCA contribution to environmental space
}
\examples{
library(humboldt)

##load environmental variables for all sites of the study area 1 (env1). Column names should be x,y,X1,X2,...,Xn)
env1<-read.delim("env1.txt",h=T,sep="\\t")

## load environmental variables for all sites of the study area 2 (env2). Column names should be x,y,X1,X2,...,Xn)
env2<-read.delim("env2.txt",h=T,sep="\\t") 

## remove NAs and make sure all variables are imported as numbers
env1<-humboldt.scrub.env(env1)
env2<-humboldt.scrub.env(env2)

##load occurrence sites for the species at study area 1 (env1). Column names should be 'sp', 'x','y'
occ.sp1<-na.exclude(read.delim("sp1.txt",h=T,sep="\\t"))

##load occurrence sites for the species at study area 2 (env2). Column names should be 'sp', 'x','y' 
occ.sp2<-na.exclude(read.delim("sp2.txt",h=T,sep="\\t"))

##convert geographic space to espace
zz<-humboldt.g2e(env1=env1, env2=env2, sp1=occ.sp1, sp2=occ.sp2, reduce.env = 2, reductype = "PCA", non.analogous.environments = "NO", env.trim.mcp = T, e.var=c(3:21),  col.env = e.var, mcp.buffer.sp1 = 200, mcp.buffer.sp2 = 200, env.pca = T, rarefy.dist = 50, rarefy.units="km", env.reso=0.41666669, kern.smooth = 1, PROJ = F,  R = 100, run.silent = F)

## plot pca contributions
humboldt.plot.contrib(zz$pca.cal$co,zz$pca.cal$eig)
}
\seealso{
\code{humboldt.sample.spp,humboldt.g2e, humboldt.equivalency.test, humboldt.background.test, humboldt.niche.overlap, humboldt.plot.niche,humboldt.doitall} which use or depend on outputs of this function
}
