% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humboldt55.R
\name{humboldt.plot.histrogram}
\alias{humboldt.plot.histrogram}
\title{Plot equivalency or background tests: simulation histogram}
\usage{
humboldt.plot.histrogram(x, type = "D", title = "equivalency")
}
\arguments{
\item{x}{Output from niche equivalency test (humboldt.equivalency.test).}

\item{type}{Values "D" or "I", which correspond to niche similarity using Schoener's D and Hellinger's I measurements}

\item{title}{Title of plot}
}
\value{
Plots the results output from the equivalency test (humboldt.equivalency.test).
}
\description{
Plot equivalency or background tests: simulation histogram
}
\examples{
library(humboldt)

##load environmental variables for all sites of the study area 1 (env1). Column names should be x,y,X1,X2,...,Xn)
env1<-na.exclude(read.delim("env1.txt",h=T,sep="\\t"))

## load environmental variables for all sites of the study area 2 (env2). Column names should be x,y,X1,X2,...,Xn)
env2<-na.exclude(read.delim("env2.txt",h=T,sep="\\t")) 

##load occurrence sites for the species at study area 1 (env1). Column names should be 'sp', 'x','y'
occ.sp1<-na.exclude(read.delim("sp1.txt",h=T,sep="\\t"))

##load occurrence sites for the species at study area 2 (env2). Column names should be 'sp', 'x','y'
occ.sp2<-na.exclude(read.delim("sp2.txt",h=T,sep="\\t"))

##convert geographic space to espace
zz<-humboldt.g2e(env1=env1, env2=env2, sp1=occ.sp1, sp2=occ.sp2, reduce.env = 2, reductype = "PCA", non.analogous.environments = "NO", env.trim= T, e.var=c(3:21),  col.env = e.var, trim.buffer.sp1 = 200, trim.buffer.sp2 = 200, env.pca = T, rarefy.dist = 50, rarefy.units="km", env.reso=0.41666669, kern.smooth = 1, PROJ = F,  R = 100, run.silent = F)

##store espace scores for sp1 and environments 1,2 and both environments combined output from humboldt.g2e
scores.env1<-zz$scores.env1[1:2]
scores.env2<-zz$scores.env2[1:2]
scores.env12<- rbind(zz$scores.env1[1:2],zz$scores.env2[1:2])
scores.sp1<-zz$scores.sp1[1:2]
scores.sp2<-zz$scores.sp2[1:2]

## run create a grid of Environmental Space Function
z1<- humboldt.grid.espace(scores.env12,scores.env1,scores.sp1,kern.smooth=1,R=100)
z2<- humboldt.grid.espace(scores.env12,scores.env2,scores.sp2,kern.smooth=1,R=100)

## run equivalency test
a<- humboldt.equivalency.test(z1,z2,rep=100,kern.smooth=1)

##plot equivalency test
humboldt.plot.histrogram(a,"D","Equivalency") 
}
\seealso{
\code{humboldt.g2e, humboldt.equivalency.test, humboldt.background.test, humboldt.niche.similarity, humboldt.plot.niche,humboldt.doitall} which use or depend on outputs of this function
}
