% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humboldt86.R
\name{humboldt.smoothing.eval}
\alias{humboldt.smoothing.eval}
\title{Visualize and select kernel smoothing parameter}
\usage{
humboldt.smoothing.eval(input.g2e, k.min = 0.25, k.max = 2, k.step = 0.25)
}
\arguments{
\item{input.g2e}{This is the output from humboldt.g2e}

\item{k.min}{The minimum k-value to evalute (default=0.25)}

\item{k.max}{The maximum k-value to evalute (default=2)}

\item{k.step}{The step value to evaluated k-values between the mmaximum and minimum (default=0.25, this means with other defaults, this function will look at the follwing k-values: 0.25, 0.5, 0.75, 1, 1.25, 1.5, 1.75, 2)}
}
\value{
This tool returns selected K-value for further use.
}
\description{
Visualize and select kernel smoothing parameter
}
\examples{
library(humboldt)

##load environmental variables for all sites of the study area 1 (env1). Column names should be x,y,X1,X2,...,Xn)
env1<-read.delim("env1.txt",h=T,sep="\t")

## load environmental variables for all sites of the study area 2 (env2). Column names should be x,y,X1,X2,...,Xn)
env2<-read.delim("env2.txt",h=T,sep="\t") 

## remove NAs and make sure all variables are imported as numbers
env1<-humboldt.scrub.env(env1)
env2<-humboldt.scrub.env(env2)

##load occurrence sites for the species at study area 1 (env1). Column names should be 'sp', 'x','y'
occ.sp1<-na.exclude(read.delim("sp1.txt",h=T,sep="\t"))

##load occurrence sites for the species at study area 2 (env2). Column names should be 'sp', 'x','y'. 
occ.sp2<-na.exclude(read.delim("sp2.txt",h=T,sep="\t"))

##perform modeling to determin imporant variables
reduc.vars<- humboldt.top.env.maxent(env1=env1,env2=env2,sp1=occ.sp1,sp2=occ.sp2,rarefy.dist=40, rarefy.units="km",env.reso=0.0833338,e.var=(3:21),pa.nu=10000,method="contrib",contrib.greater=5)

##use new variables for env1 and evn2, use as you normally would do for env1/env2 (input above)

##for example, input into converted geographic space to espace
zz<-humboldt.g2e(env1=reduc.vars$env1, env2=reduc.vars$env2...)
## select best kernel smoothing parameter based on plots
kern.smooth<-humboldt.smoothing.eval <- function(input.g2e, k.min=0.25,k.max=2,k.step=0.25)

}
