% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humboldt85.R
\name{humboldt.top.env}
\alias{humboldt.top.env}
\title{Humboldt Top Env (old function)}
\usage{
humboldt.top.env(...)
}
\description{
This is an alias for humboldt.top.env so that older scripts continue to function.  This is an alias for the function humboldt.top.env.brt.
}
